CMakeLists.txt now works
... But the mapserver CMake module is not perfect, and proposes to add
as headers locations / linking libraries, the same as the ones needed by
the package maintainer to build mapserver. That is not exactly what we
expect: We wanted the cmake provided module for mapserver to provide us
the includes path to mapserver as well as the library name to link
against (and not the whole library set needed by mapserver).

Never mind, should do the trick for now.